backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
python-multipart==0.0.6
sqlalchemy==2.0.25
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1